DB function, fix typo, add new packages.
Add a timestamp to the function that disconnects old DB connections
so that I can see when it happened.

There was a typo in the source block header for the log4j mode that
was causing the file to not tangle correctly.

Add new blocks for web-beautify and apropospriate theme.