Disable warning messages update XML functions.
The warning message for large changes to a buffer has been disbled.

Pretty print and xquery XML functions have been updated to work on
the selected region or the current buffer if a region is not
active.  This also required a change to the XML hydra as the region
xquery option is no longer valid.