Minor changes
Fix server start configuration.  Change the location of the lines
that remove the line between buffers so that it is loaded at the
same time as the theme.
Add new key binding for switching windows.
Diminish several modes that I missed.
Fix SQL functions.  The check to make sure that the SQLI process
was started was broken so the string was never sent to the process.
Unbind C-a,e,n,p,f,b to force me to use god mode properly.